doing fix code_mesh

+++ limbrary-0.1/limbrary/model_view/mesh.cpp
@@ -34,7 +34,7 @@ namespace lim
 	{
 		glBindVertexArray(vert_array);
 		glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, element_buf);
-		glDrawElements(GL_TRIANGLES, (int)(tris.size()*3), GL_UNSIGNED_INT, 0);
+		glDrawElements(GL_TRIANGLES, tris.size()*3, GL_UNSIGNED_INT, 0);
 	}
 	Mesh* Mesh::clone()
 	{
